refactor: move subcmd funcs to dedicated dir, rename `quickstart` to `populate`
i feel like "populate" better gets across what the command is doing.
