+++ test/cmd/services/root_test.go
@@ -1,17 +1,16 @@
-package test
+package services_test
 
 import (
 	"testing"
 
-	"github.com/vague2k/rummage/pkg"
-	"github.com/vague2k/rummage/pkg/database"
+	s "github.com/vague2k/rummage/cmd/services"
 	"github.com/vague2k/rummage/testutils"
 )
 
 func TestAttemptGoGet(t *testing.T) {
 	t.Run("Successfully gets go package", func(t *testing.T) {
 		test := "github.com/gorilla/mux"
-		err := pkg.AttemptGoGet(test)
+		err := s.AttemptGoGet(test)
 		if err != nil {
 			t.Errorf("Expected no error, but got an error: \n%s", err)
 		}
@@ -19,7 +18,7 @@ func TestAttemptGoGet(t *testing.T) {
 
 	t.Run("Errors when not a go package", func(t *testing.T) {
 		test := "notagopackage"
-		err := pkg.AttemptGoGet(test)
+		err := s.AttemptGoGet(test)
 		if err == nil {
 			t.Errorf("Expected error, but got nil: \n%s", err)
 		}
@@ -27,11 +26,7 @@ func TestAttemptGoGet(t *testing.T) {
 }
 
 func TestUpdateRecency(t *testing.T) {
-	tmp := t.TempDir()
-	r, err := database.Init(tmp)
-	if err != nil {
-		t.Errorf("Could not open db: \n%s", err)
-	}
+	r := testutils.DbInstance(t)
 	defer r.DB.Close()
 
 	item, err := r.AddItem("item")
@@ -39,7 +34,7 @@ func TestUpdateRecency(t *testing.T) {
 		t.Errorf("Issue adding item to db: \n%s", err)
 	}
 	t.Run("Properly increments score", func(t *testing.T) {
-		got := pkg.UpdateRecency(r, item)
+		got := s.UpdateRecency(r, item)
 		expected := 4.0
 
 		testutils.AssertEquals(t, expected, got.Score)
